+++ plugins/quay/src/components/QuayDashboardPage/index.ts
@@ -1,1 +1,0 @@
-export { QuayDashboardPage } from './QuayDashboardPage';