+++ plugins/tekton/CHANGELOG.md
@@ -1,3 +1,24 @@
+## @janus-idp/backstage-plugin-tekton [1.3.0](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-tekton@1.2.0...@janus-idp/backstage-plugin-tekton@1.3.0) (2023-04-28)
+
+
+### Features
+
+* **tekton:** tekton pipeline run list sort ([#253](https://github.com/janus-idp/backstage-plugins/issues/253)) ([356f637](https://github.com/janus-idp/backstage-plugins/commit/356f637d183eaa58e27be383db5834167f13c058))
+
+## @janus-idp/backstage-plugin-tekton [1.2.0](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-tekton@1.1.1...@janus-idp/backstage-plugin-tekton@1.2.0) (2023-04-25)
+
+
+### Features
+
+* **tekton:** add latest pipelineRun visualization ([#242](https://github.com/janus-idp/backstage-plugins/issues/242)) ([87f7f5c](https://github.com/janus-idp/backstage-plugins/commit/87f7f5c3753c100dfd50fe8e6359e22651cfb3ae))
+
+## @janus-idp/backstage-plugin-tekton [1.1.1](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-tekton@1.1.0...@janus-idp/backstage-plugin-tekton@1.1.1) (2023-04-19)
+
+
+### Bug Fixes
+
+* **tekton:** updates pf react-topology to standalone release ([#263](https://github.com/janus-idp/backstage-plugins/issues/263)) ([e739463](https://github.com/janus-idp/backstage-plugins/commit/e73946309fecfb3785419cf3543feefff4f54111))
+
 ## @janus-idp/backstage-plugin-tekton [1.1.0](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-tekton@1.0.0...@janus-idp/backstage-plugin-tekton@1.1.0) (2023-04-17)